mixins

+++ store/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import get_object_or_404
 from django.db.models import Count
 from rest_framework.decorators import api_view
+from rest_framework.mixins import ListModelMixin, CreateModelMixin
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.views import APIView
@@ -62,35 +63,27 @@ class CollectionList(APIView):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-class CollectionDetail(APIView):
-    def get(self, request, id):
-        collection = get_object_or_404(
-            Collection.objects.annotate(
-                products_count=Count('products'), pk=id)
-        )
+@api_view(['GET', 'PUT', 'DELETE'])
+def collection_detail(request, id):
+
+    collection = get_object_or_404(
+        Collection.objects.annotate(
+            products_count=Count('products'), pk=id)
+    )
+    if request.method == 'GET':
+
         serializer = COllectionSerializer(collection,)
         return Response(serializer.data)
-
-    def put(self, request, id):
-        collection = get_object_or_404(
-            Collection.objects.annotate(
-                products_count=Count('products'), pk=id)
-        )
+    elif request.method == 'PUT':
         serializer = COllectionSerializer(
             collection,
             data=request.data,
         )
         serializer.is_valid(raise_exception=True)
         serializer.save()
-        return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def delete(self, request, id):
-        collection = get_object_or_404(
-            Collection.objects.annotate(
-                products_count=Count('products'), pk=id)
-        )
+    elif request.metho == 'DELETE':
         if collection.products.count() > 0:
             return Response({'error': 'Collection cannot be deleted as it already in use'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
